+++ packages/lib/config/definitions.js
@@ -196,6 +196,12 @@ InstanceGroup.define({
 	type: "number",
 	optional: true,
 });
+InstanceGroup.define({
+	name: "auto_start",
+	description: "Automatically start this instance when the slave hosting it is started up",
+	type: "boolean",
+	initial_value: false,
+});
 InstanceGroup.finalize();
 
 /**
@@ -279,6 +285,12 @@ FactorioGroup.define({
 	description: "Synchronize banlist with master server",
 	type: "boolean",
 	initial_value: true,
+});
+FactorioGroup.define({
+	name: "max_concurrent_commands",
+	description: "Maximum number of RCON commands trasmitted in parallel",
+	type: "number",
+	initial_value: 5,
 });
 FactorioGroup.finalize();